style: Split the files and arrange
Add Files
 - InputCar.swift
 - InputAttempt.swift
